Refactor modals to components instead of pages
